requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
flask-cors==6.0.1
openai==2.6.1
python-dotenv==1.1.1
anyio==4.11.0
distro==1.9.0
httpx==0.28.1
jiter==0.11.1
pydantic==2.12.3
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.15.0
idna==3.11
httpcore==1.0.9
h11==0.16.0
annotated-types==0.7.0
pydantic-core==2.41.4
typing-inspection==0.4.2